+++ DEMToolbox/classes/particle_samples.py
@@ -1,4 +1,5 @@
 import numpy as np
+import pyvista as pv
 from .particle_attribute import ParticleAttribute
 
 class ParticleSamples():
@@ -37,6 +38,17 @@ class ParticleSamples():
         1D array of the vector 3 cell centers in the sample space
     vector_3_bounds: (np.ndarray)
         Array of the vector 3 bounds in the sample space
+
+    Methods
+    -------
+    to_vtm(filename=None):
+        Render the sampled cells as cubes in 3D space. If filename 
+        is provided, the rendered cubes will be saved as a .vtm file 
+        with the specified name. If filename is not provided, 
+        the rendered cubes will be saved as a .vtm file with the 
+        name of the samples column. vtp files are saved for each 
+        individual cube in the sample space, in a folder named after 
+        the samples column.
     """
     def __init__(self, 
                  name, 
@@ -46,10 +58,13 @@ class ParticleSamples():
                  particles, 
                  n_sampled_particles, 
                  n_unsampled_particles,
+                 vector_1=None,
                  vector_1_centers=None,
                  vector_1_bounds=None,
+                 vector_2=None,
                  vector_2_centers=None,
                  vector_2_bounds=None,
+                 vector_3=None,
                  vector_3_centers=None,
                  vector_3_bounds=None,
                  ):
@@ -63,9 +78,53 @@ class ParticleSamples():
         self.n_occupied_cells = np.size(occupied_cells)
         self.n_sampled_particles = n_sampled_particles
         self.n_unsampled_particles = n_unsampled_particles
+        self.vector_1 = vector_1
         self.vector_1_centers = vector_1_centers
         self.vector_1_bounds = vector_1_bounds
+        self.vector_2 = vector_2
         self.vector_2_centers = vector_2_centers
         self.vector_2_bounds = vector_2_bounds
+        self.vector_3 = vector_3
         self.vector_3_centers = vector_3_centers
         self.vector_3_bounds = vector_3_bounds
+
+    def to_vtm(self, filename=None):
+
+        z, y, x = np.meshgrid(self.vector_3_centers,
+                              self.vector_2_centers,
+                              self.vector_1_centers,
+                              indexing='ij'
+        )
+
+        cube_centers = np.stack([x, y, z], axis=-1).reshape(-1, 3)
+        data = []
+
+        for i, (x, y, z) in enumerate(cube_centers):
+            cube = pv.Cube(
+                center=(x, y, z),
+                x_length=self.vector_1_bounds[1] - self.vector_1_bounds[0],
+                y_length=self.vector_2_bounds[1] - self.vector_2_bounds[0],
+                z_length=self.vector_3_bounds[1] - self.vector_3_bounds[0],
+            )
+
+            cube["id"] = np.full(cube.n_cells, i)
+            data.append(cube)
+
+        blocks = pv.MultiBlock(data)
+
+        # Rotate the blocks to align with the original vectors
+        rotation_matrix = np.eye(4)
+        rotation_matrix[:3, :3] = np.array([self.vector_1, 
+                                             self.vector_2, 
+                                             self.vector_3]).T
+        blocks.transform(rotation_matrix, inplace=True)
+
+
+        # Save the blocks to a .vtm file if filename is provided, 
+        # otherwise save with the name of the samples column
+        if filename is not None:
+            blocks.save(filename)
+        else:
+            blocks.save(f"{self.name}.vtm")
+
+        return